Replace KeystreamBlockGenPage with general implementation
