+++ ChaCha/Visualization/ChaChaPresentation.xaml
@@ -90,7 +90,6 @@
           <Button Grid.Row="2" Margin="5" Width="160" Height="18.75" HorizontalAlignment="Left" Name="PageButton_StateMatrixInitialization">State Matrix Initialization</Button>
           <StackPanel Grid.Row="3" Margin="5" Orientation="Horizontal">
             <TextBlock Width="160" Height="18.75" HorizontalAlignment="Center">Keystream Block Generation:</TextBlock>
-            <Button Margin="5,0,0,0" Width="32" Height="18.75" Name="PageButton_KeystreamBlockGeneration_1">1</Button>
             <StackPanel Name="PageButtonPanel_KeystreamBlockGeneration" Orientation="Horizontal">
             </StackPanel>
           </StackPanel>
